BACKEND/requirements.txt
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-Cors==3.0.10
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
peewee==3.14.4
PyJWT==2.1.0
PyMySQL==1.0.2
python-environ==0.4.54
six==1.16.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
